Ocfs2-tests: Add multi-node testing binary for xattr_tests.
A mpi compatible program  for xattr multiple-nodes testing,should be complied/executed under
the environment of OpenMPI-1.2.5 or later.

it was based on the testcases designed for xattr-test.c,which will be executed concurrently among multiple nodes.

This common/atomic testing binary also can perform a sanity checker or a stress puncher by specifying the workload and
arguments.it was generally organized by a testing script to perform a thorough test.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>
Signed-off-by: Marcos E. Matsunaga <Marcos.Matsunaga@oracle.com>