fsck-test.sh: Check existence of LOG_DIR and set pipefail.
The script was not testing for the existence of LOG_DIR before creating the
logfile. It now checks for it and if it does not exist, it will try to create.
In case of failure, the script will abort.

Another point was that "| tee -a ${LOGFILE}" is used in many places including
a diff that determine if the testcase was successful or not. In this case, 
it is being added "set -o pipefail" so that it will carry on if the diff fails
and not signal a false success.

Signed-off-by: Marcos Matsunaga <Marcos.Matsunaga@oracle.com>

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>