Splice had a limit of 1000 bytes to write through the pipe.
I've increased the size to 10,000,000 bytes to make it more flexible.

Also, made the splice lenght to be printed to stderr instead of
stdout. This is to make the splice_read work as it was piping everything
and causing error in the splice_test.py script that is being added here.

Added the script splice_test.py

Makefile was changed to install splice_test.py

Signed-off-by: Marcos Matsunaga <Marcos.Matsunaga@oracle.com>

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>