ocfs2-test: add nocheckdirs option to the program.
During the execution of run_buildkernel.py, it checks for each node's 
directory size on all directories specified. This is a long and expense
process if the test has already run once on the same nodes and directories.
This option will skip this check, but the user must make sure the
directories are in good condition. 

This fix is useful in cases where a crash happens during the build process
and needs to be restarted.

Signed-off-by: Marcos Matsunaga <Marcos.Matsunaga@oracle.com>

Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>