ocfs2-test: tunefs-test.sh Fixed syntax & adapted to new debugfs.ocfs2 output.
Other than the syntax error, the new output of the debugfs.ocfs2 required 
a change to the script to prevent false error reported. It also need to
have a change to the way some flags were compared as there were not that
many flags when this script was first written. It now checks for the 
existence of "backup-super" instead of the flag "1" as before.
It also looks for "inline-data" instead of "InlineData".

Signed-off-by: Marcos Matsunaga <Marcos.Matsunaga@oracle.com>

Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>