Ocfs2-test: Improve resize_test.sh for tunefs.ocfs2 changes.
As the output of new tunefs.ocfs2 changed when doing resizing(Now it
output nothing even if we turn the vebose mode on), our test tools
therefore need to be updated accordingly. from my point of view, it
would be better to both check the return code of the testing cmd and
output of debugfs.ocfs2 to see if the volume size grows correctly.

Some other improvements also included in the patches for permission and
path problems.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>
Signed-off-by: Tao Ma <tao.ma@oracle.com>