Ocfs2-test: Let multi-inline-tests accept different device name.
As we know, a same volume may be treated by different device name among nodes who share
the disk. and our former multi-inline-tests will definitely fail due to this reason.

To solve this issue, we try to pass UUID of a volume to mpi binary, and the binary
convert the UUID to a device name on its own, which will be later called by ocfs2_open() in
our testing binary.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>
Signed-off-by: Marcos E. Matsunaga <Marcos.Matsunaga@oracle.com>