Ocfs2-test: Add destructive testcase for reflink_test v3.
Per sunil's request, we're going to add a destructive testcase for reflink
to expose an expected issue in existing reflink kernel codes: O_DIRECT writes
will not flush the metadata accordingly when being performed on reflinked files
after the completion of write operation, which means a reflinked file may wrongly
points to an old reflinked extent after a unexpectedly crash of machine.

The v3 testcase was also a somewhat quick&dirty one which incorporate sunil's
suggestions, and add checksum verification for data chunk, we also do the verification
for reflinks in this version.

After doing reflinks, we fork procs to perform random writes on a reflinked file,
and each write is going to be logged accordingly via wire to a remote listener server.
then crash the machine somehow, afterwards, the logfile recorded by listener will be
used for verification.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>