Ocfs2-test: Handling boundary case of verfiy_tests on reflink_test.
For the newly added destrutive tests(intentional to reproduce bug 1191),
we verify all chunks in original and reflinked file after destructive reboot
by glancing writes record in a remote logfile. So the filesize we specify in
the destructive and verify test should be the same as expected.

This patch handle the case when the filesize were not the same, it avoids a
short read or segfault when such extreme situation happened.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>
Acked-by: Tao Ma <tao.ma@oracle.com>