Resize_test: Kill a deadloop case in resize_test.sh
Currently, resize_test will be running into a deadloop if our partition size
didn't align with clustersize when growing the volumesize to the end of the
partition, in that case, ${blocks} will never be equal to ${partsz}, which caused
while loop endless.

This patch attempts to kill this by checking the above case to break the loop.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>
Signed-off-by: Marcos E. Matsunaga <Marcos.Matsunaga@oracle.com>