Discontig-bg-test: Add utility to generate file with desired extents.
We really need a tool to generate files with massive extents, by specifying
filesize and chunksize(chunksize here actually is the size for each extent).
As a result, we'll be able to populate a desired number of extent as expected.

It could be later used as a stress extents consumer which eats up the extents
from extent_alloc sharply.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>