Discontig-bg-test: Add a tool to fill up contiguous block group.
The original intention for this script is to let it become multi-nodes-aware,
otherwise we may not bother ourselves to separate this job from the script of
discontig-activator. when we're executing multi-nodes discontig tests, openmpi
call this script among nodes, and do the contiguous-bg fillup job respectively
by calculating their own slot numbers. that way, it becomes quite simple to get
block allocaters of all nodes filled up.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>