Ocfs2-test: Add directio test for ocfs2-test.
We were lacking of a dedicated directio test in ocfs2-test to perform
an isolated and comprehensive test for O_DIRECT ONLY, the idea here
is originally driven by the intention to write a destructive testcase
of O_DIRECT append writes, where bug 1285 was exposed:

http://oss.oracle.com/bugzilla/show_bug.cgi?id=1285

Thus, we may have a good chance here to introduce a common and thorough
O_DIRECT test for the sake of our testing coverage.

The test is going to torture and verify four kinds of O_DIRECT writes:

   - O_DIRECT writes within i_size
   - O_DIRECT writes outside i_size
   - O_DIRECT writes within hole
   - O_DIRECT in destructive mode.

A following verification can be launched the validate writes by consulting
a dedicated local or remote log where all writes get recorded.

The test will always be ready to grow to absorb more sophosticated testcases.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>