Ocfs2-test: Implement testing body.
It filles the testing bodies for all testing modes.

The test uses the number of concurrent processes and file size
to tune the workload for writes intensity.

The testing file is expected to consist of CHUNKs, a CHUNK is
treated as a basic unit to perform write/read/verification with
a unique pattern being filled, while the CHUNK_SIZE here also
needs to be 512-bytes aligned.

Like the 'fill_verify_holes', directio test also recoreds all
writes as tuples in a logfile for following verification, more
specifically, destructive test uses a remote listener for log
writing.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>