Ocfs2-test: Add manipulation codes for file verification, and make it public.
The original intention to make it public was driven by the requirement of doing
a thorough verification test for defragmentation after crash.

File verification is very common and critical in many area of filesystem's validation
test, especially to validate the file's data after a crash, which makes it necessary
to keep a series of verification codes into libo2test library to benifit all tests,
such as directio/reflink tests, and our destructive tests for deframentation later
also can take advantage from this.

The essence of verification logic is to format file's data in terms of chunks, with a
certain pattern for each chunk, each chunk consists of following parts:

         * chunkno
	 * timestamp
	 * checksum
	 * random chars
         * checksum
	 * timestamp
	 * chunkno

During the tests, it also requires the writers to write file in chunks, with a give pattern.
writing record can be kept locally, or remotely via socket.

After crash, the verification test can be done, with a remote/local write log being provided.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>