defrag-test: add an utility to frag fs on purpose.
Actually mark's reservation patches is making the attempt to frag fs no that
esay in a short team, the utility is try to growing-write the fs in small chunks
with massive processes being spawned and interleaved, to simulate the situation
of a long-lived fs. then helps fraging the files/fs in a relatively tolerable time.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>