discontig: use local alloc when active discontig block group
Lots of small files will be allocated when active discontig block
group, if not use local alloc, the data will be allocated from
global bitmap, since the chain in global bitmap are almost full and
with the nearly the same free bitmap at that time, so for every cluster,
the chain with most free bitmap will be used, this can cause a case
like this, cluster A allocated from chain X, and next cluster A+1
allocated from chain Y, so these small files are full of fragments.
This is not wanted, use local alloc to avoid this.

Signed-off-by: Junxiao Bi <junxiao.bi@oracle.com>
Reviewed-by: Eric Ren <zren@suse.com>