multiple_run: automatically detect the IP interface
In the past, the default ip interface name is eth0,
this is not flexible, and in the latest Linux distributions the
IP interface is renamed to ens[n].
So, I add a f_detect_if() function to detect the IP interface.
Secondly, add a missed parameter in flock_unit test case.